re-push
mistake while merging last resulted in loss of some changes, they were now fixed and back in the code

+++ src/main/java/dk/dtu/cdio2/PlayerManager.java
@@ -24,7 +24,7 @@ public class PlayerManager {
             this.accountID = am.createAccount().getAccountID();
         }
 
-        public void withdrawMoney(int m) {
+        public void withdrawMoney(double m) {
             am.getAccount(this.accountID).withdraw(m); //Mangler at connecte am med mp her
         }
